Full nullable support added to the typechecker.  Note that it will be weakened in the future.  The following code is illegal:
String s;
s = "thing";

Dataflow analysis will allow this kind of thing in the future.

"var" keyword for automatic type changed to "auto".  Support added to typechecker, but only for local variables.  Automatic types for fields opens up the possibility of circularly unresolvable types and is surprisingly difficult to typecheck.