Fixes to the typechecker with respect to the new calling convention for new.
We may want to re-evaluate this, since it does make things more cumbersome.

Right now, there is a restriction of only a single new on the left side with no methods that can be called off the right side of the invocation.

Syntax for inner class instantiation is:

Outer o = new Outer();
Outer:Inner i = new o:Inner();