Changed "construct" and "array" to "create", which now does all allocation except for singletons ("instance").
Change explicit create invocation back to this() or super() (removed the "create" from this:create() and super:create()).

Changed "noop" to "skip" for empty statements.

Removed TypeArgument, since it was just an alias for Type.

Created Destroy() production in grammar, so that destroy can be called on an object.  No type-checking for it yet.