Added typechecker support for Number, allowing arbitrary types to be added, subtracted, multiplied, divided, and modded.
Support updated for CanIndex, allowing arbitrary types to use array style subscripts.

CanIterate is working, but foreach loops do not work properly since the type declaration and scope is not working right.  It's next on the list to fix.

Non-nullable non-primitive variables will no longer need an initializer to typecheck.  This functionality needs to be put into the TAC, but it is making writing useful tests annoying right now.