Added foreach typechecking support.
Changed imports to automatically include all files in all directories (even when not the main file being compiled).

Changed typechecking structure so that there is no separate parsing phase.  Instead, the type checker returns the node of the file of interest.  Otherwise, typechecking code is duplicated to treat the parsed file differently from other files that need to be parsed, which results in mostly duplicated code.