Changed definition of CanIndex so that the return value is always nullable.  Updated LinkedList to follow this definition.
Simplified SequenceType checking for method parameters and return values but also made it more robust.  (Appropriate nullable checking and a few other things are covered.)

Added typechecking so that Array<int> and int[] are considered equivalent in the typechecker.  Note:  There is no notion of dimensions in the Array<int> formulation.  The backend should check for incompatibilities.