Changed type checker to use CanAdd<T>, CanSubtract<T>, ... CanModulus<T> for operator overloading.
Bitwise behavior uses similar interfaces.

Currently, CanNegate<T> is required for a unary negation or (meaningless) unary plus.  My implementation does not allow unary negation on unsigned types.  This is logical, but is it reasonable?

CanEqual<T> is needed for == and !=.

Nodes which perform operations (such as ASTAdditiveExpression) are now OperationNodes, which have a list of MethodSignatures inside of them.  For consistency, even some nodes which can only have a single operator are still OperationNodes.

Minor fix for some generic issues.

Typechecking code still needs some cleanup.

Note the following change:

class Outer<T>
{
	public class Inner<T>
	{}
}

This code currently creates a new type parameter T for Inner (in addition to the T for the outer class).  Is this useful?  Should we prevent the same type parameter from being used.  In older typechecker versions, this code did not fully create a new type parameter.