Continued improvement of immutable and readonly semantics.
Now, it is possible to write a readonly method that does not change anything inside the object but can also return a mutable reference to an object.

The typechecker keeps track of which things are readonly only for the duration of the method and which things must be readonly (or immutable) on return, using the hidden modifiers RETURN_READONLY and RETURN_IMMUTABLE.

From a design perspective, would it be better to mark things as temporarily readonly instead of mapping return readonly to readonly and then replacing it later?