Removed redundant circular extends and implements check from TypeUpdater, using only the check in TypeInstantiater.
Added enforcement of immutable methods, meaning that mutable parent methods must be overridden in immutable child classes.

For simplicity, updated the copy() method in Object to be readonly, so that all immutable classes will not have to override it.

Since the freeze() method is always implemented as native and non-overridable (because of immutability), a customized version of freeze() is now added to every class.