Finally unified generic interfaces.
Fixed Windows.ll and Linux.ll to reflect changes in Object needed to make File work.

Updated Main.ll so that unhandled exceptions are caught at the highest level.

Added OutOfMemoryException for situations when malloc() returns NULL, but it doesn't seem to work.  Perhaps the problem is that there is already no memory left to allocate the OutOfMemoryException?

The sorting example seems to run out of memory for merge sort, probably because of a small heap size and because there is no garbage collection.

The critical .equals() method in Type was renamed to .typeEquals() so that .equals() can be used only for hash tables.

However, ArrayList is now broken for reasons related to generic types being either indistinguishable or incorrectly seen as different.