Added support for switches.
They now work for integer types and Strings.  Multiple values mapping to the same code is done by separating the values with commas in the case statement.  The code below each case statement must be surrounded with braces.

The break keyword is not used for switches.  Instead, each block jumps to the end of the switch after completion.