Removed error and view keywords from the compiler for upcoming release.
It's likely that error doesn't have a lot of good use cases, and it's
not clear how views will work at this point.

Added full support and test cases for the assert keyword.

Also, changed singleton classes to allow them to inherit from a regular
class (other than Object).  Not sure if that's a good idea.