Changed check so that it does not require a try-recover.
With a try-recover, check will go to the recover block on null.
Otherwise, it will throw an UnexpectedNullException.

Updated configuration parsing so that arch works correctly.  Without
explicitly specifying arch="32", Windows will use Unwind64.ll, breaking
exception handling.

Fixed Windows error output so that it can be redirected to a file.
Windows documentation seems to suggest that problems may happen with
wide-characters.

Added functionality to OutputTest.java so that it can expect specific
output for standard error as well as standard output.