Reduced redundant typechecking errors.
Adding a typechecker error to the internal list of errors now allows a
list of types to be part of the adding process.  If any of the types are
Type.UNKNOWN, a special error type, then the error will not be added to
the list.  An AST node should only have Type.UNKNOWN as its type if
there was a previous error.

Many fewer repetitive errors should be reported, though some may have
slipped through.  This system also does not stop multiple reports if,
for example, the same undeclared identifier is used on multiple lines.