Changed configuration so that link and other commands work.
Instead of using the Apache Digester class, JAXB is used to
automatically create a Shadow class based on the configuration.xsd XML
schema file.  It may be possible to clean up this approach later since
the Configuration class duplicates much functionality from the Shadow
class.

Also fixed bug that stopped 128 from being used as a byte literal, even
with a negation in front of it.