Changed grammar so that array initializers can appear anywhere, not just on the right hand side of an array declaration.
Note: There is some danger of { thing.getX() } being interpreted as an
array initializer instead of a block, assuming that { thing.getX(); }
was intended.

It may be necessary to add some kind of type marker to initializers:
int[]{ 3, 5, 7, 8}.  Doing so would reduce confusion with blocks as well
as allow int[,] for multidimensional arrays, since the current approach
treats nested array initializers strictly as arrays of arrays.