Added run-time check that will throw an InterfaceCreateException if a user attempts to create an interface.
A normal interface would be caught much earlier in the typechecker.
Even an interface that includes the appropriate create() method would be
caught at compile time.  However, a generic type parameter that happened
to be an interface that included a create() method that was created
inside of the generic object was a loophole until this fix.

Added unit tests and updated the creation of the generics file to
support generics with interface types (whose method lists are null).