Fix so that fields are initialized in the order they were written (not the order that they are laid out within the object).
This change is important when fields are dependent on the prior
initialization of other fields.  Note that this still doesn't not catch
the error when fields access each other out of order.  The data flow
analysis isn't to that point yet.

Added small tests and changes to the grammar so that a newline inside of
a String literal gives a better error message.  Also, use of ++ and --
now gets a special message.