Simplified the Class class so that it only has size() and width() methods.  The sizeValue() is no longer needed.
Note that size() will give the actual size of the underlying class for
reference types.  It gives the size of the wrapper object for primitive
types.  It gives the dimensions of the array for array types.

The width() method gives the size needed per object when stored in an
array.  For reference types, this is the size of a pointer.  For
primitive types, it's their actual size.  For arrays and interfaces,
it's the space they need.