Updated generic class creation so that fully instantiated generic classes whose types are known in the file can use the named generic class instead of reconstructing it at run time.
This change is an optmization that allows most generic classes to be
used directly instead of constructed.  The exception is if code in
something like ArrayList<T> creates a HashSet<T>.  Since T is unknown at
compile time, HashSet<T> will need to be constructed at run-time.

In order to make these changes, the compilation process parses the
beginnings of .LL files, where the generic classes are listed.  Then,
all known generic classes are added to the set of classes before the
main() method is called.  Unfortunately, the approach is ugly and
fragile in the face of changes to .LL file generation.