Updated constant generation so that inner class constants can be defined in terms of outer class constants.
Note that outer class constants cannot be defined in terms of inner
class constants.  Also, constants of two separate inner classes cannot
be defined in terms of each other.

Also added a fix to simplify constant accesses, removing the need to get
a "this" reference.

Also fixed toString() for classes with no package specified, putting the
@ sign after the default keyword.