Updated type collector to trigger recompilation of files that are dependent on files that have changes.
To do so, the type collector (potentially) makes two passes, discarding
the type information from the first one but using it to make a list of
types that require recompilation.

The process is conservative, sometimes recompiling things that do not
need to be recompiled.  For example, a single file that never ends up
getting compiled will trigger the recompilation of everything in the
directory (since there is no .ll file newer than the .meta file).