Added warning when field is never used.
Since fields are always private, a field that is never used inside of
the class is never used.  Special care is needed to check inner classes.

As an unrelated fix, I removed the field list inside of Type, which
duplicated the storage of the field map but allowed the fields to be
retrieved in the order they were declared.  This functionality is now
accomplished by using a LinkedHashMap for the field map.