Changed TACTree so that each tree has a TACOperand representing the value computed by the TACTree.
This change makes the code easier to think about.  Previously, appending
one tree to another returned a TACOperand if that was the last thing in
the tree.  Doing so required awkward type checking to see if the last
thing was a TACOperand and also required the creation of a TACNodeRef
for those situations when the value computed by a tree is not the last
thing in the tree (usually because the last thing is now a label).

Unfortunately, this change does require the TACBuilder to explicitly set
the TACOperand for a TACTree when one is generated.  I believe we're
gaining some clarity at the cost of a little simplicity.