Implemented two-step parsing in ANTLR.
ANTLR 4 is very powerful, but it is somewhat slow.  Its designers
created tools that allow users to use the faster ANTLR 3-style parsing
and then fall back on ANTLR 4-style if a file can't be parsed.  By using
this two-step parsing, it significantly increases the speed of the
Shadow compiler.

For reference, using purely ANTLR 4-style parsing takes over 800 seconds
to run all tests, but using two-step parsing reduces the time to only
579 seconds.