First steps toward garbage collection.
After turning loads and stores of local variables into phi nodes, we
have to turn those phi nodes back into loads and stores, but only for
reference variables.  The @llvm.gcwrite calls that we're going to be
using to do reference counting require real memory cells (not ephemeral
phi node stuff).

Thus, all of these reference variables require a memory location to be
allocated on the stack with alloca.