Added basic comparer classes for the List<T>
testEqualityComparer currently fails the test because of broken
generics/interfaces. When those tests pass, we can move on to using the
comparer with the List, LinkedList, and what not. This will avoid the
need of having List<T is CanEqual<T>> and give the developer more
flexibility to pass classes with no CanEqual interface.