A few modifications to the C micro-framework and other minor changes
There are now two versions to the extract the actual pointer from the
ShadowPointer: ExtractRawPointer() and ExtractPointer().
- ExtractRawPointer(ShadowPointer): takes a ShadowPointer type, which is
the Shadow-initialized pointer, and returns the unmanaged pointer, as a
void*. This requires the developer to cast the pointer:
(ActualType*)ExtractRawPointer(ref).
- ExtractPointer(Type, ShadowPointer): Equivalent to the method above,
but returns the desired type so no need for casting:
ExtractPointer(ActualType, ref), return ActualType*.

Mutexes should be working now, but unfortunately the signaler isn't
working yet. There are a few issues with concurrency, and I might need
to work the Signaler from the ground up, both Shadow and C codes.