Removed SLF4J and moved back to Log4j 2 completely.
Initial change to SLF4J was because of what turned out to be unrelated
errors in the Shadow plugin.  SLF4J is a wrapper that allows many
different logging frameworks to be used, and that level of flexibility
is probably not needed for Shadow.  Log4j 2 works fine and is a
stand-alone logging solution.