More updates to the C micro-framework
- Fixed a few issues with the C compiler which caused the compiled
assembly files to not be added to the linking process when they are
already compiled.
- Added two defines SHADOW_ARCH32 and SHADOW_ARCH64 which specify
whether the operating system we're on is 32 or 64 bit. No need for weird
defines such __WIN32 and __WIN64; those defines work across all
platforms and can be used by including <ShadowCore.h>.
- Commented out the tests in OutputTest.java which used to be
implemented in the previous threading commits, and as the old threading
features are added, the tests will be uncommented.