Fixed deep generic ClassSet bug.
Generic class sets were not being initialized before the main method's
object's constructor was called.  As a consequence, a constructor that
created a generic object could, in some cases, try to ask the class set
for data when the class set wasn't initialized.

Now, the code that initializes the ClassSet objects happens very early
in the main template code in Main.ll and NoArguments.ll.