Fixed FileTest.shadow.
Previous test opened file for writing and then started reading. Our
current File code tries to close the file and then reopen for both
reading and writing...but it's not clear to me what the behavior should
be in that case.  Explicitly closing the file fixes the test.

This test (and File.shadow) should be re-examined.