Made all inner classes "static" inner classes in the grammar and typechecker.
Inner classes are no longer linked to an object of outer class type.

The annoying part about this is that type parameters must be "repeated"
for inner classes that depend on the type parameters.

This change is designed to simplify the internals of the compiler as
well as to prevent unnecessary references from inner class objects to
outer class objects.

Although these references seem harmless, they are a breeding ground for
circular references, an unwise practice for a language with
reference-counting garbage collection.