Changed type parameter mangling to be more robust.
In the past, all type parameters were mangled to T for simplicity.
Although this works in most situations, it ran into trouble with methods
that can mangle the same.

Imagine a class Wombat with two type parameters, U and V.  It could have
methods:

public method(U input) => () and
public method(V input) => ()

which would both mangle to the same name.  Now, one will be
Wombat_Mmethod_TU and the other will be Wombat_Mmethod_TV.  The T is
prepended to the type parameter name in case someone named a type
parameter A.  A is used in mangling to mark an array type.