Store cast of primitive values to Object into a temporary variable.
Doing so allows the object (a wrapper allocated specifically to hold the
primitive) to be garbage collected when its reference count falls to
zero.

Note that this commit also undoes the previous commit which removed the
storage of null to all of the local variables at the beginning of a
method.  There are cases when, for example, a variable is stored to if a
certain condition is true but not otherwise.  In those situations, the
variable must be initialized to null so that when garbage collection is
attempted, the variable is null (which has a special case to deal with
it) rather than filled with junk data.