Improved garbage collection.
As an optimization, stores of variables do not decrement the old value
if it has never been stored to before.  However, the compiler was too
generous with deciding which stores had never been stored before.  This
has been changed to be more accurate.

Results on DrMemory are cleaner than ever before, showing no leaks on
most test cases.