Changed -r (--human-readable) option to generate LLVM files first.
The previous system used a special `OutputStream` class that duplicated
all LLVM output, sending it to both a .ll file and into the stdin of the
LLVM `opt` program.  It turns out that doing so is very slow.  Making
the .ll file first and then optimizing it with `opt` is much faster.