First steps towards decorators and threading
- Decorator pattern is under development, and is the main way now to
import and export methods.
- Only compiler decorators are currently supported, and are:
   - ImportNative (imports method from .native.ll)
   - ImportAssembly (imports method from .c)
   - ImportMethod (imports method from another class)
   
   - ExportAssembly (exports a method to be callable from C)
   - ExportMethod (exports a method to be importable from the class)
- Decorators are in the form of @(Type1, Type2, ...) where TypeN derives
from the Decorator interface.
- A method decorator needs to derive from MethodDecorator interface.
- Only method decorators are currently supported, but decorators would
in theory expand to classes too. 
- To write a decorator that is a hint to the compiler, we would derive
from CompilerMetaDecorator interface. All the above decorators derive
from it.
- Decorators could be useful in modifying a certain method's behavior,
for example, we could have a LogDecorator, that would log a method's
behavior, such as the time it took to execute, or the parameters passed
to it and the result the method returned. This could be especially
useful for quick testing.


Note1: I'm currently working on threading in parallel to this, so soon enough, changes would come with more decorators, and working Threads.
Note2: Decorators would change later to @(Type1(...params), Type2(...params), ...) so in theory, a method could automatically create the decorators with the desired parameters, and in turn execute them as needed, in the order they were assigned.
Note3: Compiler decorators are not integrated with methods, and are only placeholders. They are not initialized nor appear in the generated Shadow code.
Note4: Decorators will not appear in .meta files as they are considered additional code, and would not appear just as method contents don't appear in the .meta.