Improved method references so that they can be passed as parameters.
All the machinery was about making it so that type-checking was happy.
To make it go smoothly, I added a MethodReferenceType which is
essentially a wrapper for a MethodType.  However, these two types are
used at different types.  A MethodReferenceType is turned by the
compiler into a Method object.  It is the type used to represent method
references.

The MethodType is used as a general way of representing the types of
methods and is stored in the MethodSignature class as well as other
places.