Updated tests to JUnit 5, including Maven requirements.
Unfortunately, test suites are not supported by JUnit 4, so the tests
themselves are JUnit 5, but the code that runs all the tests as a suite
is JUnit 4, requiring a separate dependency.

Future JUnit 5 versions are supposed to support suites, so we can
hopefully simplify tests when those features are available.