Changed TAC generation and LLVM output to use new EH syntax
This syntax uses catchswitch/catchpad/catchret/cleanuppad/cleanupret.

Unfortunately, it doesn't work because finally blocks are called both
when an exception is unwinding and normally.  The cleanuppad semantics
require a token, and LLVM isn't happy if a cleanuppad jumps to a block
of code and jumps back to a cleanupret when normal code can jump to the
same block and back.

This problem might be solved if the token type was allowed in a phi, but
it isn't.  The semantics require a funclet bundle for methods called
inside of a cleanuppad or a catchpad.

Ultimately, this means that code from finally blocks will need to have
two copies in the TAC and LLVM: unwinding and normal execution.

Could this mean an exponential blow-up of code if a finally is inside of
a finally which is inside of a finally?