Update that creates Windows-based exception handling (EH) mechanisms.
Note: Although some of the changes simplify the code base for a future
return to Linux/Mac EH, EH is broken on those platforms by this update.

Unfortunately, differences in EH mean that the initial plan of Shadow to
produce identical LLVM IR for all platforms is impossible.

Also, funclet bundles need to be specified for method calls, which this
update doesn't do yet.