Shadow now links with Clang under Windows.
Exceptions (SEH) are not yet working because we need to fix the
catchpads.  Nevertheless, we are close.

Note that it was necessary to use LLVM 10 to compile, since LLVM 9
seemed to have a bug related to math intrinsics in Windows.

We also fixed the version compare tool so that it can compare complex
version numbers.