Windows exception handling finally works.
To model the funclet requirements of MSVC SEH, it was necessary to
create special functions that are called whenever a finally is entered.
These functions have the benefit of producing single copies of finally
blocks.  Previously, finally blocks were duplicated in code: once for
execution and once when unwinding from an exception.  Failure to have
this duplication might have been one of the causes of previous issues
with Linux exception handling.

Unfortunately, the code to do control flow analysis is significantly
complicated with these special functions, which are treated as if they
live inside of the principle method.

Linux exception handling was broken by fixing Windows.  That fix and
other clean-up is essential.