Fixed generics with type bounds.
Now, type parameters will be treated like their class bound (if one
exists) rather than Object.

Also, there's small fix in this push that makes sure that parents of
classes are included in the set of used types.  Otherwise, it's possible
for LLVM to refer to a type that is undefined.