Changed Strings to contain an array of ubyte values
Deep down, this change is insignificant to LLVM, but it is slightly more
honest from the Shadow perspective.  All character values should be
thought of as sequences of non-negative bytes (ubytes).  Since ubytes
are unsigned, they can also be considered subtypes of codes, allowing
for assignment without explicit casting.

I hope this change went in the right direction, because it caused all
kinds of crazy, unexpected bugs when I first made the change.