Fixed Windows bug when there are spaces in the path
In Windows (and perhaps in other OSes), a bug occurred when Shadow was
installed in a directory whose path had spaces because of bad URL to URI
translation.

This bug was particularly pernicious because it occurred only when
running code from inside a JAR and thus did not show up during unit
testing.