Changed compiler configuration files from XML to JSON
We have been fighting with JAXB for years, and no one's excited about XML these days anyway.  Few of the projects are in active development, and JAXB support was dropped from core Java with Java 11+.  This commit changes configuration files over to JSON, using the Jackson library for parsing.

Shadow configuration files are quite simple anyway.  A nice side benefit of using Jackson is that it also supports YAML, so YAML support for configuration files would be easy to add on or migrate to in the future.

Also, parent configuration files were always part of Shadow configuration files but never implemented correctly.  In theory, they now should be.

Final side benefit: Lots of Maven dependencies go away, and the size of the final Shadow JAR is smaller, only 5MB right now.