Added lookup in SHADOW_HOME
If a suitable config file (shadow.json) is not found elsewhere, the system will look in the directory specified by the SHADOW_HOME environment variable.

This step was necessary for an installation on Linux to work where the user installing Shadow does not have sudo privileges.

A fundamental problem with the current Shadow compiler is that there is no easy way to specify source directories and binary directories.  Consequently, the standard Shadow library must be stored in user directories, not in the shared space.