Huge painful reorganization to allow support for separate src and bin directories.
In fact, unlimited pairs of src and bin directories are possible.  The configuration JSON file allows pairs to be specified in the import field.  Likewise, the system field is a list that should contain exactly three elements for src, include, and bin for the system files.

This approach is intended to make a more flexible compiler and a system for cleaner installations, since the system src and bin can be in read-only folders while local src and bin directories allow for user files.

This commit also removes the opt phase, which I believe is redundant, running optimizations that are done by llc anyway.