Changes to standardize conventions for import/export naming
Methods implemented in C should always start with two underscores (__).

Methods implemented in Shadow but exported for use in C should start with a single underscore (_).

To avoid naming conflicts, after the initial underscores, methods should be named with prefixes that specify the classes they are included with.  Underscores are used to replace dots (.), and two underscores are used to replace the at sign (@).