The simplest threaded programs work.
It's possible to spawn a thread, perform some action, and then join the thread.

The system requires more testing, but Mutex and ConditionVariable seem to work as well. It seems clear that Thread objects should *not* be deeply copied, since their functionality depends on being shared across threads. Perhaps that's true aout Mutex and ConditionVariable objects as well.