Small fixes to compiler error-checking
Previously, the compiler would allow a property that returns a nullable value to be used as if it was nullable. Now, the nullability constraints on properties are enforced.

Also, it was possible to crash the parser with certain kinds of malformed input. Now, an extra null check should give a useful error message in those situations.