Removed ExportMethod as an attribute
The classes that needed ExportMethod were associated with Thread. To make this work, it was necessary to make those classes inner classes of Thread, including the CurrentThread singleton (now Thread:Current). Doing so simplifies the compiler and collapses a lot of thread mechanics into a single file (good for locality but makes a large file).

While a good move overall, there are still errors, perhaps because of bad coding in Thread.