Fixed issues with ConditionVariable
In Windows, ConditionVariable uses a CriticalSection that can be entered multiple times by the same thread. However, a CriticalSection should not be entered multiple times if it's going to be used for a Windows ConditionVariable. These changes try to enforce ownership of the CriticalSection by a particular thread (which is only allowed to enter it once). There are some concerns about setting the owner non-atomically that might need to be addressed in the future.