Simplified spawn syntax
Now, a spawn() will take an object of any class that implements CanRun. This object could be created by performing a :create() right inside the spawn(), or it could be a pre-existing object.

At present, neither types nor method references can be spawned, but both of these are possibilities for the future.