Fixed another bug in Windows file I/O and in String parsing
Now, a broken pipe error is treated like an EOF. This makes redirected input work as expected.

Also added an EndOfFileException.shadow, which is thrown by TextInput when trying to read a line or String (or int or double) past the end of a file.

Finally, added NumberFormatException errors when trying to parse a String into a number, if the String is empty.